extract export tests
