+++ test/e2e/utils/surveyExport/extract.js
@@ -1,0 +1,9 @@
+import fs from 'fs'
+
+import { extractZip } from '@server/utils/file/fileZip'
+
+export const extractZipFileAndCheck = async ({ zipPath, extractedPath, surveyExtractedPath }) => {
+  await extractZip(zipPath, extractedPath)
+  await expect(fs.existsSync(extractedPath)).toBeTruthy()
+  await expect(fs.existsSync(surveyExtractedPath)).toBeTruthy()
+}
